+++ python/src/biosimulant_model_compatibility_standard/__init__.py
@@ -30,4 +30,4 @@ __all__ = [
     "validate_object",
 ]
 
-__version__ = "0.1.0a5"
+__version__ = "0.0.1"